+++ include/sm2d/colliders.h
@@ -32,14 +32,13 @@ enum ColliderType
 struct Collider
 {
     ColliderType type;
-    union
-    {
-        ColAABB    aabb;
-        ColCircle  circle;
-        ColPolygon polygon;
-    };
-    Rigidbody* body;
-    int        treeIndex = -1; // Index in the AABB tree
+    ColAABB      aabb;
+    ColCircle    circle;
+    ColPolygon   polygon;
+    Rigidbody*   body = nullptr;
+    int          treeIndex = -1; // Index in the AABB tree
+
+    int bodyIndex = 0; // Ignore this, it's only for the editor
 
     Collider(ColliderType type, const ColAABB& aabb, Rigidbody* body)
        : type(type), aabb(aabb), body(body)
@@ -55,9 +54,10 @@ struct Collider
     }
 
     ~Collider() {} // This is just here so the compiler doesn't yell at me
+    Collider() {}
 };
 
-struct Manifold 
+struct Manifold
 {
     bool      colliding;        // Are they colliding?
     glm::vec2 collisionNormal;  // Direction of the collision used for impulse calculation