+++ include/sm2d/systems.cpp
@@ -9,6 +9,22 @@
 namespace sm2d
 {
 
+void RigidbodyStartSys()
+{
+    for (EntityID ent : SceneView<Rigidbody>(engineState.scene))
+    {
+        auto rigid = engineState.scene.Get<Rigidbody>(ent);
+
+        if (rigid->transform == nullptr)
+        {
+            if (auto trans = engineState.scene.Get<Transform>(ent))
+            {
+                rigid->transform = trans;
+            }
+        }
+    }
+}
+
 void RigidbodySys()
 {
     for (EntityID ent : SceneView<Rigidbody>(engineState.scene))
@@ -45,11 +61,28 @@ void RigidbodySys()
     }
 }
 
+// PLEASE TURN THIS OFF WHEN YOU CAN, THIS FUNCTION IS AN ABSOLUTE CATASTROPHE
 void DebugSys()
 {
     for (EntityID ent : SceneView<Collider>(engineState.scene))
     {
         auto collider = engineState.scene.Get<Collider>(ent);
+
+        if (collider->body == nullptr)
+        {
+            if (auto rigid = engineState.scene.Get<Rigidbody>(ent))
+            {
+                collider->body = rigid;
+
+                if (rigid->transform == nullptr)
+                {
+                    if (auto trans = engineState.scene.Get<Transform>(ent))
+                    {
+                        rigid->transform = trans;
+                    }
+                }
+            }
+        }
 
         if (collider->type == ColliderType::sm2d_AABB)
         {
@@ -76,6 +109,7 @@ void DebugSys()
         }
         else if (collider->type == ColliderType::sm2d_Polygon)
         {
+            UpdatePolygon(*collider);
             std::vector<glm::vec3> threedpoints;
             for (auto& point : collider->polygon.worldPoints)
             {
@@ -128,6 +162,14 @@ void ColliderSys()
     {
         auto collider = engineState.scene.Get<Collider>(ent);
 
+        if (collider->body == nullptr)
+        {
+            if (auto bod = engineState.scene.Get<Rigidbody>(ent))
+            {
+                collider->body = bod;
+            }
+        }
+
         if (collider->body->type == BodyType::sm2d_Static || !collider->body->awake)
         {
             continue;
@@ -157,9 +199,10 @@ void ColliderSys()
 }
 
 REGISTER_START_SYSTEM(ColliderStartSys);
+REGISTER_START_SYSTEM(RigidbodyStartSys);
 
-// REGISTER_SYSTEM(DebugSys);
 REGISTER_SYSTEM(RigidbodySys);
 REGISTER_SYSTEM(ColliderSys);
+REGISTER_EDITOR_SYSTEM(DebugSys);
 
 } // namespace sm2d