+++ src/main.cpp
@@ -10,68 +10,19 @@ Camera camera(glm::vec3(0.0f, 0.0f, 3.0f), glm::vec3(0.0f, 1.0f, 0.0f), -90.0f, 
 
 int main(int argc, char** argv)
 {
-    Window window("Prism", SCR_WIDTH, SCR_HEIGHT);
+    Window window("Prism", SCR_WIDTH, SCR_HEIGHT, false, true);
     // glfwSwapInterval(1);
 
-    unsigned int groundTex = Utils::LoadTexture("res/textures/background.png");
-    unsigned int slugariusTex = Utils::LoadTexture("res/textures/Slugarius.png");
-
     Scene scene;
-
-    EntityID ground = scene.AddEntity();
-    scene.AssignParam<Transform>(ground, glm::vec3(0.0f, -1.0f, 0.0f), glm::vec3(0.0f, 0.0f, 0.0f),
-                                 glm::vec3(10.0f, 1.0f, 1.0f));
-    scene.AssignParam<SpriteRenderer>(ground, groundTex, glm::vec4(1.0f, 1.0f, 1.0f, 1.0f),
-                                      "Ground");
-    scene.AssignParam<sm2d::Rigidbody>(ground, sm2d::BodyType::sm2d_Static,
-                                       scene.Get<Transform>(ground), 100.0f, false, 0.7f, 0.7f,
-                                       0.0f, true, 1.5f);
-    scene.AssignParam<sm2d::Collider>(
-        ground, sm2d::ColliderType::sm2d_Polygon,
-        sm2d::ColPolygon({glm::vec2(-5.0f, -0.5f), glm::vec2(-5.0f, 0.5f), glm::vec2(5.0f, 0.5f),
-                          glm::vec2(5.0f, -0.5f)}),
-        scene.Get<sm2d::Rigidbody>(ground));
-    scene.AssignParam<Name>(ground, "Ground");
-
-    EntityID sprite = scene.AddEntity();
-    scene.AssignParam<Transform>(sprite, glm::vec3(0.0f, 1.0f, 0.0f), glm::vec3(0.0f, 0.0f, 0.0f),
-                                 glm::vec3(1.0f, 1.0f, 1.0f));
-    scene.AssignParam<SpriteRenderer>(sprite, slugariusTex, glm::vec4(1.0f, 1.0f, 1.0f, 1.0f),
-                                      "Sprite");
-    scene.AssignParam<sm2d::Rigidbody>(sprite, sm2d::BodyType::sm2d_Dynamic,
-                                       scene.Get<Transform>(sprite), 2.0f, true, 0.56f, 0.56f, 0.0f,
-                                       false, 0.4f);
-    scene.AssignParam<sm2d::Collider>(
-        sprite, sm2d::ColliderType::sm2d_Polygon,
-        sm2d::ColPolygon({glm::vec2(-0.5f, -0.5f), glm::vec2(-0.5f, 0.5f), glm::vec2(0.5f, 0.5f),
-                          glm::vec2(0.5f, -0.5f)}),
-        scene.Get<sm2d::Rigidbody>(sprite));
-    scene.AssignParam<Name>(sprite, "Sprite");
-
-    EntityID box = scene.AddEntity();
-    scene.AssignParam<Transform>(box, glm::vec3(2.0f, 2.0f, 0.0f), glm::vec3(0.0f, 0.0f, 0.0f),
-                                 glm::vec3(1.0f, 3.0f, 1.0f));
-    scene.AssignParam<SpriteRenderer>(box, slugariusTex, glm::vec4(1.0f, 1.0f, 1.0f, 1.0f),
-                                      "Circle");
-    scene.AssignParam<sm2d::Rigidbody>(box, sm2d::BodyType::sm2d_Dynamic, scene.Get<Transform>(box),
-                                       10.0f, true, 0.56f, 0.56f, 0.0f, false, 2.5f);
-    scene.AssignParam<sm2d::Collider>(
-        box, sm2d::ColliderType::sm2d_Polygon,
-        sm2d::ColPolygon({glm::vec2(-0.5f, -1.5f), glm::vec2(-0.5f, 1.5f), glm::vec2(0.5f, 1.5f),
-                          glm::vec2(0.5f, -1.5f)}),
-        scene.Get<sm2d::Rigidbody>(box));
-    scene.AssignParam<Name>(box, "Box");
 
     engineState.SetScene(scene);
     engineState.SetCamera(camera);
 
     Renderer::Init(false, true);
 
-    StartStartSystems();
+    StartEditorStartSystems();
 
     ImGuiLayer::Init();
-
-    sm2d::Collider* col2 = engineState.scene.Get<sm2d::Collider>(sprite);
 
     std::vector<sm2d::Manifold> colResults;
 
@@ -81,7 +32,12 @@ int main(int argc, char** argv)
     {
         // Start of frame
         ImGuiLayer::NewFrame();
-        UpdateSystems();
+
+        UpdateEditorSystems();
+        if (playing)
+        {
+            UpdateSystems();
+        }
 
         // Main loop logic
         // ---
@@ -99,7 +55,6 @@ int main(int argc, char** argv)
         window.Update();
     }
 
-    DestroyPhysics();
     ImGuiLayer::Terminate();
 
     return 0;