+++ src/systems.cpp
@@ -24,7 +24,8 @@ void MeshRendererSys()
         glClear(GL_DEPTH_BUFFER_BIT);
 
         for (unsigned int i = 0; i < 6; ++i)
-            Renderer::depthShader.setMat4("shadowMatrices[" + std::to_string(i) + "]", light.shadowTransforms[i]);
+            Renderer::depthShader.setMat4("shadowMatrices[" + std::to_string(i) + "]",
+                                          light.shadowTransforms[i]);
         Renderer::depthShader.setFloat("farPlane", light.radius);
         Renderer::depthShader.setVec3("lightPos", light.position);
 
@@ -55,8 +56,7 @@ void SpriteRendererSys()
 {
     for (EntityID ent : SceneView<SpriteRenderer>(engineState.scene))
     {
-        Renderer::RenderSprite(ent, engineState.projMat,
-                               engineState.camera->GetViewMatrix());
+        Renderer::RenderSprite(ent, engineState.projMat, engineState.camera->GetViewMatrix());
     }
 }
 
@@ -73,7 +73,7 @@ void RigidBody3DStartSys()
             continue;
         }
 
-        JPH::Vec3 transPosition(trans->position.x, trans->position.y, trans->position.z);
+        JPH::Vec3  transPosition(trans->position.x, trans->position.y, trans->position.z);
         JPH::RVec3 RtransPosition = transPosition;
 
         if (rigid->colliderType == ColliderType::Box)
@@ -85,26 +85,29 @@ void RigidBody3DStartSys()
 
             // Create the shape
             JPH::ShapeSettings::ShapeResult floor_shape_result = floor_shape_settings.Create();
-            JPH::ShapeRefC floor_shape = floor_shape_result.Get(); // We don't expect an error here, but you can check
-                                                                   // floor_shape_result for HasError() / GetError()
+            JPH::ShapeRefC                  floor_shape =
+                floor_shape_result.Get(); // We don't expect an error here, but you can check
+                                          // floor_shape_result for HasError() / GetError()
 
             // Step 1: Convert Euler angles to GLM quaternion
             glm::quat glmQuat = glm::quat(trans->rotation);
 
             // Step 2: Convert GLM quaternion to Jolt quaternion
-            JPH::Quat joltQuat(glmQuat.x, glmQuat.y, glmQuat.z, glmQuat.w); // Jolt uses (x, y, z, w) order
+            JPH::Quat joltQuat(glmQuat.x, glmQuat.y, glmQuat.z,
+                               glmQuat.w); // Jolt uses (x, y, z, w) order
 
-            // Create the settings for the body itself. Note that here you can also set other properties like the
-            // restitution / friction.
+            // Create the settings for the body itself. Note that here you can also set other
+            // properties like the restitution / friction.
             JPH::BodyCreationSettings floor_settings(
                 floor_shape, RtransPosition, joltQuat,
-                rigid->state == BodyState::Dynamic ? JPH::EMotionType::Dynamic : JPH::EMotionType::Static,
+                rigid->state == BodyState::Dynamic ? JPH::EMotionType::Dynamic
+                                                   : JPH::EMotionType::Static,
                 rigid->state == BodyState::Dynamic ? Layers::MOVING : Layers::NON_MOVING);
 
             floor_settings.mCollisionGroup.SetGroupID(rigid->groupID);
             // Create the actual rigid body
-            rigid->body =
-                bodyInterface.CreateBody(floor_settings); // Note that if we run out of bodies this can return nullptr
+            rigid->body = bodyInterface.CreateBody(
+                floor_settings); // Note that if we run out of bodies this can return nullptr
 
             // Add it to the world
             if (rigid->body != nullptr)
@@ -117,7 +120,8 @@ void RigidBody3DStartSys()
             float radius = rigid->capsuleRadius;
             float height = rigid->capsuleHeight;
 
-            // Capsule is defined by its half height (distance between the centers of the hemispheres)
+            // Capsule is defined by its half height (distance between the centers of the
+            // hemispheres)
             JPH::CapsuleShapeSettings capsule_shape_settings(height * 0.5f, radius);
             capsule_shape_settings.SetEmbedded();
 
@@ -125,15 +129,17 @@ void RigidBody3DStartSys()
             glm::quat glmQuat = glm::quat(trans->rotation);
 
             // Step 2: Convert GLM quaternion to Jolt quaternion
-            JPH::Quat joltQuat(glmQuat.x, glmQuat.y, glmQuat.z, glmQuat.w); // Jolt uses (x, y, z, w) order
+            JPH::Quat joltQuat(glmQuat.x, glmQuat.y, glmQuat.z,
+                               glmQuat.w); // Jolt uses (x, y, z, w) order
 
             // Create the shape
             JPH::ShapeSettings::ShapeResult capsule_shape_result = capsule_shape_settings.Create();
-            JPH::ShapeRefC capsule_shape = capsule_shape_result.Get();
+            JPH::ShapeRefC                  capsule_shape = capsule_shape_result.Get();
 
             JPH::BodyCreationSettings capsule_settings(
                 capsule_shape, RtransPosition, joltQuat,
-                rigid->state == BodyState::Dynamic ? JPH::EMotionType::Dynamic : JPH::EMotionType::Static,
+                rigid->state == BodyState::Dynamic ? JPH::EMotionType::Dynamic
+                                                   : JPH::EMotionType::Static,
                 rigid->state == BodyState::Dynamic ? Layers::MOVING : Layers::NON_MOVING);
 
             capsule_settings.mCollisionGroup.SetGroupID(rigid->groupID);
@@ -156,11 +162,12 @@ void RigidBody3DStartSys()
 
             // Create the shape
             JPH::ShapeSettings::ShapeResult sphere_shape_result = sphere_shape_settings.Create();
-            JPH::ShapeRefC sphere_shape = sphere_shape_result.Get();
+            JPH::ShapeRefC                  sphere_shape = sphere_shape_result.Get();
 
             JPH::BodyCreationSettings sphere_settings(
                 sphere_shape, RtransPosition, JPH::Quat::sIdentity(),
-                rigid->state == BodyState::Dynamic ? JPH::EMotionType::Dynamic : JPH::EMotionType::Static,
+                rigid->state == BodyState::Dynamic ? JPH::EMotionType::Dynamic
+                                                   : JPH::EMotionType::Static,
                 rigid->state == BodyState::Dynamic ? Layers::MOVING : Layers::NON_MOVING);
 
             sphere_settings.mCollisionGroup.SetGroupID(rigid->groupID);
@@ -200,12 +207,13 @@ void RigidBody3DSys()
         float z = rotationOfSphere.GetZ();
         float w = rotationOfSphere.GetW();
         // Sync box position and rotation with Jolt Physics
-        engineState.scene.Get<Transform>(ent)->position =
-            glm::vec3(positionOfSphere.GetX() + rigid->offset.x, positionOfSphere.GetY() + rigid->offset.y,
-                      positionOfSphere.GetZ() + rigid->offset.z);
+        engineState.scene.Get<Transform>(ent)->position = glm::vec3(
+            positionOfSphere.GetX() + rigid->offset.x, positionOfSphere.GetY() + rigid->offset.y,
+            positionOfSphere.GetZ() + rigid->offset.z);
 
         glm::quat quatRotation(-z, y, x, w);
-        glm::vec3 eulerRotation = glm::eulerAngles(quatRotation); // Converts quaternion to Euler angles
+        glm::vec3 eulerRotation =
+            glm::eulerAngles(quatRotation); // Converts quaternion to Euler angles
 
         engineState.scene.Get<Transform>(ent)->rotation = eulerRotation;
     }
@@ -221,21 +229,27 @@ void LightStartSys()
 
         light->shadowTransforms.clear();
 
-        glm::mat4 shadowProj =
-            glm::perspective(glm::radians(90.0f), (float)SHADOW_WIDTH / (float)SHADOW_HEIGHT, 1.0f, 25.0f);
+        glm::mat4 shadowProj = glm::perspective(
+            glm::radians(90.0f), (float)SHADOW_WIDTH / (float)SHADOW_HEIGHT, 1.0f, 25.0f);
 
         light->shadowTransforms.push_back(
-            shadowProj * glm::lookAt(lightPos, lightPos + glm::vec3(1.0f, 0.0f, 0.0f), glm::vec3(0.0f, -1.0f, 0.0f)));
+            shadowProj * glm::lookAt(lightPos, lightPos + glm::vec3(1.0f, 0.0f, 0.0f),
+                                     glm::vec3(0.0f, -1.0f, 0.0f)));
         light->shadowTransforms.push_back(
-            shadowProj * glm::lookAt(lightPos, lightPos + glm::vec3(-1.0f, 0.0f, 0.0f), glm::vec3(0.0f, -1.0f, 0.0f)));
+            shadowProj * glm::lookAt(lightPos, lightPos + glm::vec3(-1.0f, 0.0f, 0.0f),
+                                     glm::vec3(0.0f, -1.0f, 0.0f)));
         light->shadowTransforms.push_back(
-            shadowProj * glm::lookAt(lightPos, lightPos + glm::vec3(0.0f, 1.0f, 0.0f), glm::vec3(0.0f, 0.0f, 1.0f)));
+            shadowProj * glm::lookAt(lightPos, lightPos + glm::vec3(0.0f, 1.0f, 0.0f),
+                                     glm::vec3(0.0f, 0.0f, 1.0f)));
         light->shadowTransforms.push_back(
-            shadowProj * glm::lookAt(lightPos, lightPos + glm::vec3(0.0f, -1.0f, 0.0f), glm::vec3(0.0f, 0.0f, -1.0f)));
+            shadowProj * glm::lookAt(lightPos, lightPos + glm::vec3(0.0f, -1.0f, 0.0f),
+                                     glm::vec3(0.0f, 0.0f, -1.0f)));
         light->shadowTransforms.push_back(
-            shadowProj * glm::lookAt(lightPos, lightPos + glm::vec3(0.0f, 0.0f, 1.0f), glm::vec3(0.0f, -1.0f, 0.0f)));
+            shadowProj * glm::lookAt(lightPos, lightPos + glm::vec3(0.0f, 0.0f, 1.0f),
+                                     glm::vec3(0.0f, -1.0f, 0.0f)));
         light->shadowTransforms.push_back(
-            shadowProj * glm::lookAt(lightPos, lightPos + glm::vec3(0.0f, 0.0f, -1.0f), glm::vec3(0.0f, -1.0f, 0.0f)));
+            shadowProj * glm::lookAt(lightPos, lightPos + glm::vec3(0.0f, 0.0f, -1.0f),
+                                     glm::vec3(0.0f, -1.0f, 0.0f)));
 
         glGenFramebuffers(1, &light->depthMapFBO);
 
@@ -245,8 +259,8 @@ void LightStartSys()
 
         for (unsigned int i = 0; i < 6; ++i)
         {
-            glTexImage2D(GL_TEXTURE_CUBE_MAP_POSITIVE_X + i, 0, GL_DEPTH_COMPONENT, SHADOW_WIDTH, SHADOW_HEIGHT, 0,
-                         GL_DEPTH_COMPONENT, GL_FLOAT, NULL);
+            glTexImage2D(GL_TEXTURE_CUBE_MAP_POSITIVE_X + i, 0, GL_DEPTH_COMPONENT, SHADOW_WIDTH,
+                         SHADOW_HEIGHT, 0, GL_DEPTH_COMPONENT, GL_FLOAT, NULL);
         }
 
         glTexParameteri(GL_TEXTURE_CUBE_MAP, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
@@ -280,6 +294,7 @@ void AnimatorStartSys()
 
         animator->currentTime = 0.0f;
         animator->boneMatrices.reserve(200);
+        animator->model = &engineState.scene.Get<MeshRenderer>(ent)->model;
 
         for (int i = 0; i < 200; i++) animator->boneMatrices.push_back(glm::mat4(1.0f));
     }
@@ -308,16 +323,16 @@ void DeltaTimeSystem()
 }
 
 // Start systems
-REGISTER_START_SYSTEM(RigidBody3DStartSys);
-REGISTER_START_SYSTEM(LightStartSys);
-REGISTER_START_SYSTEM(AnimatorStartSys);
+REGISTER_START_SYSTEM        (RigidBody3DStartSys);
+REGISTER_START_SYSTEM        (LightStartSys);
+REGISTER_EDITOR_START_SYSTEM (AnimatorStartSys);
 
 // Regular systems
-REGISTER_SYSTEM(DeltaTimeSystem);
-REGISTER_SYSTEM(AnimatorSys);
-REGISTER_SYSTEM(MeshRendererSys);
-REGISTER_SYSTEM(SpriteRendererSys);
-REGISTER_SYSTEM(RigidBody3DSys);
-REGISTER_SYSTEM(ParticleSystemSys);
-REGISTER_SYSTEM(ButtonSys);
-REGISTER_SYSTEM(TextSys);
+REGISTER_EDITOR_SYSTEM (DeltaTimeSystem);
+REGISTER_SYSTEM        (AnimatorSys);
+REGISTER_EDITOR_SYSTEM (MeshRendererSys);
+REGISTER_EDITOR_SYSTEM (SpriteRendererSys);
+REGISTER_SYSTEM        (RigidBody3DSys);
+REGISTER_EDITOR_SYSTEM (ParticleSystemSys);
+REGISTER_EDITOR_SYSTEM (ButtonSys);
+REGISTER_EDITOR_SYSTEM (TextSys);